Convert TabGroup to its own isolated component
